Updated Contact in referance to TelephoneNumber
Now checks for string.Empty from the Number property rather than a null,
as this was throwing errors when the Telephone number was initialized,
but the values weren't complete.
Also move the check for e-mail parts to within the check for emptiness
to keep the exception from throwing on empty strings.