Added company/building list to add contact form.
Upon loading, the form queries the database to get a list of every
company and building present. It uses those to fill the list boxes on
this form.
Textboxes above each list can be used to filter the items in the list by
way of a Linq query, which also takes into account items which are
present in the matching selected listbox.