Enabled buttons to move listbox items
Added MoveListBoxItem to the buttons on the tab control. Based on the
owner of the sender and the text on the button, the items will be
shifted from one list box to another. Items in the right hand listbox
will be used when the Commit is called to associate that contact with
the particular company or building.