Change to Company List SQL Query
Made the Query a Left Join, so that Companies without any buildings
associated with them would still appear in the list. They are considered
"Not Active".
Also Changed form to only try and select a building from the "Other
Building" listbox if there are in fact any buildings present. And if
not, to populate the fields with a blank building in order to clear any
existing information from the display.