Rewrite of Company/Building Lists
Instead of constantly running LINQ queries in the MainForm class for
each of the dictionaries, we've moved all of the sorting and selection
functions to a separate class called "DatabaseList". This reduces the
amount of code required to update the listboxes on the main form or
select an item from those lists and should provide a way of maintaining
lists for use among all forms, rather than having each form perform it's
own queries.