Bug Fix & Exception Handling in Query class
The Database Table "Inspection" had modified a column outside of the
program - "InspectionType" became "Type". This caused an exception to be
thrown in the GetTable Method of Query. The Exception was thrown before
the form loaded, so it was handled by the form without alerting the
user, and furthermore allowed the SQLConnection to remain open, which
then failed the very next time the connection was tried.
Fix: a "Finally" was added to the Try..Catch in GetTable to close the
table regardless of what happens, and corrected the explicit references
to  InspectionType. Also made the Exception a bit more clear to assist
with Debugging.