Fix for Crash on Startup
In the event that no filters were enabled at startup (neither companies
from DC or MD are shown), the form bails out at the point in the load
method where the first item in the listbox is selected, which would then
trigger the information to load. We now check to see that there are
items in the listbox, and if not, do not select an item programatically.
Loading of the form would then be deferred until a checkbox item is
modified.