Changes to Update Queries
Update Queries were failing because both Update builder and the
SQLColumn were adding in single quotes. That responsability should lie
with the SQL Column. Additionally, strings that are passed to the
SQLColumn(string, string) constructor will be marked as NULL if the
value is either null or an empty string.