Added Utility for searching for reports
Accidentally deleted the data in the Reports column of the database, so
this was written to repair that. Scans the report folder and attempts to
pair them with inspection reports in the database. A little rough -
keeping it for later use but probably not going to be included in the
final