Started work on Backups
Created a helper class that forms a backup file from the SQL database.
"##" denotes a comment and "$" a table name. The first line after the
table name is a list of columns and value types, in SQL format and
separated by "|". Rows are listed after - "|" between each value and
line breaks between rows.