Added catches for SQL Duplicates
Provides catches surrounding the add company form in main form to alert
the user in the event that the insert SQL statement fails as a direct
result of a duplicate key or constraint within the database. Provides a
message box, then selects the duplicate value from the listbox and
populates the form.