Changes to Building and Inspection to reflect SQL
Added an inspectiontypes table to the database, replacing the
InspectionType column from Inspection with an integer ID called IType_ID
that refers to the new table. Made changes to the Building and
Inspection classes to reflect the change:

Inspection Constructor now checks to see if the static Inspection Types
dictionary is initialized, and if not, creates it based on the Table.
Removed the Type enum and replaced it's use with the inspectionType
dictionary