Removed Empty Class and expanded PDF Library
Removed Add Object class, as it did nothing, and expanded PDF__Document
to allow adding text and images. Still not clear on how to fit this all
in a class structure - as is, I have to make sure to call ClosePDF()
when I'm done. It doesn't appear to work through a deconstructor, but
I'm sure there's a way to get it to perform automatically.