Added check for directory when adding inspection
When adding an inspection, will now check to be sure the folders in the
path we've created for the file all exist - i.e. the inspector has a
folder. If not, asks the user if they want to create one, and if so,
creates one and places the file within.