Changes to BaseObject
Replaced the ArrayList for edits with a List<> - not sure exactly why I
went with that in the first place. The made changes to the base
constructor for BaseObject_Edited. We now check to make sure that the
item we are editing has not been changed already and if it has, we just
update the value rather than adding a whole new edit. Prevents
situations where forms change the value of an object based on this class
multiple times and we ended up with a new edit entry for each time it
was updated.