Re-write of AddInspections
Re-wrote AddInspections almost in total. Rather than create and populate
the inspection classes when it comes time to submit to the server, A
list of inspections is kept as a private variable. In the event we are
editing an inspection, it is loaded at initialization,, otherwise it is
refresehed any time the building is changed. As elements of the
inspection are edited, the inspections in the list are updated after the
values are validated. Then, when it is time to submit, the inspections
are committed in turn.