Can now edit elevator via right click on MainForm
Right clicking on the elevator list on the main form brings up the
AddElevatorForm, but it will be populated with the data from the
selected elevator. Modified the Elevator form in the same manner as the
Inspection form to maintain the ability to add a new elevator or edit an
existing depending on how the form is called.