Corrected Cert Creation
Had an issue with sizing the background image properly, along with an
oversight regarding deleting existing files of the same name. Also
changed the font size of text to 10pt and modified the number of cert
number shown on the first line to depend on the length of the formatted
list it would create.