Fixed Error with Editing Inspections
The Constructor used to edit inspections was calling a base constructor
used to add new inspections. As a result of the way new inspections were
created, the inspection List was populated with two blank inspections
before we had even gotten to this point. Quick fix is to just reset the
inspection list to empty, and fill it again with the inspections we get
from the query.