Corrected the Database List Queries
Updated the Database List Queries to work with the schema in the Azure
Database. The first one looks complicated, but all it does is Check to
see if a Company has buildings associated with it in both DC and
Maryland, and if any of those buildings are active. These are used to
quickly filter the listboxes and retrieve IDs for selected items, rather
than storing the complete list on the user's computer.