Updated Inspection Class to new schema
Also included some changes regarding the class's storage of Inspection
Type, Inspector name, and status. Inspection Type and Name are now
stored as strings. The whole ID and Linq query from a static list is too
complicated for what it accomplishes. And now that there are just three
possible values for Status, I've added a static method to convert the
string to a nullable bool. Finally, included query and possible insert
for report files that do not already exist in the database.