Corrected Error in Inspection Commit
When I was looking for a description to put into the document entry, I
was trying to get the building address based off of the inspection ID.
For new inspections, however, this will be null. It makes more sense to
grab the address using the elevator ID in any event.