Fixed Zip codes when starting with a 0
There are some zip codes which begin with a 0. As the database stores
the zip code field as an integer, they appear to have less than 5
digits. I've just padded the string it reads from the database under the
assumption that we'll never have an exception to the five digit rule.