Fixed numeric-string conversion in Telephone struct
Might be a dumb way to do it, but converting the telephone number from a
string to a 3 part number was causing it to lop off any 0's that
appeared in the beginning of either the second or third part of a number
(i.e. 410-012-3456 became 410-12-3456) Corrected it by padding the
string properties and methods to the correct length with 0's.