Added preliminary viewer for Contact Log Entries
Created a new class derived from DatabaseObject. Currently, can only be
loaded from the database via static methods GetCompanyContactLog and
GetBuildingContactLog, to prevent the SQL queries from being stored
outside of the class. May look at modifying the other classes to
something similar.
The Entries are viewed on the main form on their own tab control, and
can be restricted to show only entries related to a particular building
or to all entries from that company.